+++ projetoBackEnd/src/main/java/emakersProjetoBackEnd/service/PessoaService.java
@@ -13,6 +13,8 @@ import emakersProjetoBackEnd.data.dto.response.AdminPessoaResponseDTO;
 import emakersProjetoBackEnd.data.entity.Pessoa;
 import emakersProjetoBackEnd.exceptions.authentication.InvalidRegisterException;
 import emakersProjetoBackEnd.exceptions.general.EntityNotFoundException;
+import emakersProjetoBackEnd.exceptions.pessoas.InvalidPessoaDeletionException;
+import emakersProjetoBackEnd.repository.EmprestimoRepository;
 import emakersProjetoBackEnd.repository.PessoaRepository;
 
 
@@ -24,6 +26,9 @@ public class PessoaService {
 
     @Autowired
     private CepService cepService;
+
+    @Autowired
+    private EmprestimoRepository emprestimoRepository;
 
 
     //método que retorna todas as pessoas
@@ -39,15 +44,18 @@ public class PessoaService {
         return new AdminPessoaResponseDTO(pessoa);
     }
 
-    //método qeu adiciona uma nova pessoa
+    //método que adiciona uma nova pessoa
     public AdminPessoaResponseDTO createPessoa(AdminPessoaRequestDTO adminPessoaRequestDTO){
         if(pessoaRepository.findByEmail(adminPessoaRequestDTO.email()) != null ){
-            throw new InvalidRegisterException();
-        }else{
+            throw new InvalidRegisterException("Email já cadastrado no sistema");
+        }
+        
+        if(pessoaRepository.findByCpf(adminPessoaRequestDTO.cpf()) != null ){
+            throw new InvalidRegisterException("CPF já cadastrado no sistema");
+        }
             Pessoa pessoa = new Pessoa(adminPessoaRequestDTO);
             
             //senha padrão
-
             pessoa.setSenha("111");
             String encryptedPassword = new BCryptPasswordEncoder().encode(pessoa.getSenha());
             pessoa.setSenha(encryptedPassword);
@@ -60,13 +68,19 @@ public class PessoaService {
             pessoa.setLocalidade(cepResponseDTO.localidade());
             pessoaRepository.save(pessoa);
             return new AdminPessoaResponseDTO(pessoa);
-        }
     }
 
     //método que atualiza as informações de uma pessoa
     public AdminPessoaResponseDTO updatePessoa(AdminPessoaRequestDTO adminPessoaRequestDTO, Long idPessoa){
         Pessoa pessoa = findPessoa(idPessoa);
 
+        if(pessoaRepository.existsByEmail(adminPessoaRequestDTO.email())){
+            throw new InvalidRegisterException("Email já cadastrado no sistema");
+        }
+
+        if(pessoaRepository.existsByCpf(adminPessoaRequestDTO.cpf())){
+            throw new InvalidRegisterException("CPF já cadastrado no sistema");
+        }
 
         pessoa.setName(adminPessoaRequestDTO.name());
         pessoa.setCpf(adminPessoaRequestDTO.cpf());
@@ -88,7 +102,13 @@ public class PessoaService {
     //método que deleta uma pessoa
     public String deletePessoa(Long idPessoa){
         Pessoa pessoa = findPessoa(idPessoa);
+        if(emprestimoRepository.existsByPessoaAndStatusTrue(pessoa)){
+            throw new InvalidPessoaDeletionException(pessoa.getIdPessoa());
+        }
+
+        emprestimoRepository.deleteAllByPessoa(pessoa);
         pessoaRepository.delete(pessoa);
+
         return "Pessoa: " + idPessoa + " has been deleted";
     }
